feat: fanout log streaming
